Move useEntityData out of the tiptap SmartLink node module
